Add work-in-progress of source code
Currently summary parsing works pretty well (though I'm sure we'll
iterate on it). We still need to run it through a few tests that check
to see how it does with GitBook and mdBook compatible files. Over all
though, we can now start to work on the rendering of the content.