Change how configuration is applied
Prior to this change, preprocessors and renderers never had access to
their configuration. It would have had to be accessed via the
configured application itself. This means that there are some
redundancies, and the fact that the configuration dictionaries aren't
immutable means that it is possible to introduce bugs where there's
drift in the preprocessor or renderer from what is in the app. However,
all of this aside, having the configuration actually live closer to the
particular worker, makes a lot more sense.