Expand rendering behavior
This gets us closer to actually writing out the marked up text to
files. It is really rough right now and we still need to add tests as
well as clean up the code. However, since this is very much a WIP
draft of the code, I'm trying to evaluate  the behavior via the
command line first.