Add context to rendering hooks and use Jinja templates
We're actually rendering the converted Markdown to files! Huge
milestone. We're going to continue iterating on this and make it
better, more legible, etc. but we're close to meeting the minimum
threshold of utility for Hon. We still need to render the summary and
actually work out navigation, as well as generate a render context per
page. Some things we're going to need to think about.