Update summary structure
We're moving away from the rigid: affix, numbered, and suffix sections
to be more flexible. A summary should support an arbitrary number of
sections. The end result of this is that it should be compatible with
both mdBook and Gitbook.

Currently, these modifications are just structural and we haven't
realized the necessary changes with the parser. That comes next.