Add multi-pass template rendering
Before rendering the final result of the markdown to HTML conversion to
the outputted files, we want to render the preprocessed content as a
template first. This covers situations where macros may have been
defined and other jinja2-specific features that aren't available in
Hon.

The basic idea is that the result of preprocessing the markdown files
might still result in text that is in-and-of-itself a valid template.
Resolving this intermediate template first, and then the actual theme's
template allows us to leverage the power of jinja2 fully.