Implement a draft of the abstract parsing interface
This provides stubs for ASCIIDoc and rST, as well as a stub for the
Mistune Markdown library. This is all in service of providing more
flexibility to Hon by allowing the ability to take and convert from
more sources than just Markdown. In the case of Markdown there are two
libraries we're currently evaluating as the default Markdown parser and
we needed a way to quickly and easily swap between the two without
requiring us to change a whole host of code.

The interface provides us this flexibility.

This first pass is mostly just migrating the existing override to
Markdown Python into the new interface and then continuing with
development initiatives to further enhance the experience.