Add early support for PDF rendering
With Weasyprint we found a really great library for supporting the
creation of PDF files from HTML. Importantly, the library also supports
a number of features that I was not able to figure out how to do with
GitBook (e.g. multi-column layouts). Although none of our tests at this
point in time play around with these features and this is still a very
early implementation of the renderer it can output a PDF!