Correctly render epub files
We also moved around some of the template stuff to make our approach to
the themes a little more sensible. Initially our theming assumed that
we'd share resources, but the stylesheets for web, PDF, and ebooks are
all drastically different. We can take some direction from the web, but
we won't reuse spectre.css for non-web outputs.