Move preprocessing within the renderer
This is a pretty big shift from where we started. The initial idea was
that all summary parsing, chapter loading, and preprocessing of the
chapters happened once per book. It didn't matter whether you were
rendering it to one output, e.g. a website, or 100 different outputs.
This never felt quite right. What if there was a reason for a
preprocessor to do something specifically related to one renderer, but
not another? What if the preprocessor needed access to the rendering
context, such as with variables, to add something to it? Some of these
things were implemented with hacky work arounds, but in exploring new
ways to leverage Jinja2 as part of the preprocessing step for include
and import syntax, it became clear that we needed to move the
preprocess phase inside the renderer.

This also meant moving where the chapters got loaded, and where the
state of the chapters was held.

Chapters are also now loaded as part of the rendering phase, which
means that with both the preprocessing and the chapter loading being
done within the renderer, we're duplicating some work for ourselves.
The trade off is an isolation of the content, such that every operation
as part of the rendering pipeline is idempotent. There are no bizarre
side effects because you already rendered a book once for epub and
you're using the same in-memory chapter representation for HTML.