Change authors to author in epub templates
We shifted from authors to author recently to keep more party with
Gitbook, and after giving it some thought, the most common case is for
there to only be a single author  of a book, so the singular
configuration makes more sense. This still supports an array of authors
though, if you really want to specify more than one author for a book
as discrete entities.