Replace the include preprocessor with jinja
The include preprocessor served its purpose as a starting point for a
lot of  Hon's functionality. However, after maturing a little bit, it
became clear that we could do everything that we currently supported in
Hon's include preprocessor with Jinja. This also allowed us to support
import statements where the imported file was not necessarily in the
same directory as the current chapter.

Now, we can leverage the full power of Jinja with our include and
import statements.

The include preprocessor might be revived at some point in time in the
future as a "playpen" preprocessor (or similar code sandbox idea).