GH-7: Associated default logging handler with root (#14)
By associating the default logging handler (stdout) to the root logger,
we prevent other code from instantiating a root logger using
logging.basicConfig with an unknown format. This also prevents the
instances of multiple loggers that we've been seeing.

Specifically, this was noticed when working on the development of the
search (lunr) plugin, because the Lunr package for Python calls
logging.basicConfig.