Improve python workflow
Instead of doing everything through the make init (which would be
ideal) because we're trying to enforce the usage of Black right now as
part of the development environment using Pipenv, we're exporting the
dependencies to a requirements.txt file during the dependency
installation workflow.  I'm starting to look into poetry to see if this
gives us more flexibility.