Add infrastructure to exercise Github Actions locally using Docker
We can pretty closely emulate what Github Actions is doing by using
Docker and a shell script, this allows us to test our workflow some
before we commit it. That way we can be a little more confidant when we
make a change to our workflow whether it will actually succeed or not.